build: fix warnings with gcc 8.2.1 [#2999]
Fail to compile with gcc 8.2.1. Many warnings are due to snprintf truncation
issues.

Others are logd crash due to invalidating of iterators while clearing map, and
fixing uninitialized variable in smf.

+++ src/imm/immnd/immnd_proc.c
@@ -1887,7 +1887,7 @@ static int immnd_forkPbe(IMMND_CB *cb)
 	char execPath[1024];
 	char dbFilePath[1024];
 	int pid = (-1);
-	int execDirLen = (int)(strlen(cb->mProgName) - strlen(base));
+	size_t execDirLen = strlen(cb->mProgName) - strlen(base);
 	int dirLen = (int)strlen(cb->mDir);
 	int pbeLen = (int)strlen(cb->mPbeFile);
 	int i, j;
